+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Version 0.3.0
 - Rename package from `com.github.antag99.retinazer` to `com.darkyen.retinazer`, since the two libraries have
     diverged enough to not be substitutable any more
+- `Mapper.get` no longer returns `null` on missing component, use `Mapper.getOrNull` if you want that
 
 # Version 0.2.6 (released 2020-12-13)
 - Add `Engine.getServices` to retrieve service subclasses